refactor: restructured the datatable component
