+++ src/lib/components/DataTable/RowCount.svelte
